misc: refacto pagination
