+++ src/components/ListMovies.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState, useMemo } from "react";
-import ReactPaginate from "react-paginate";
 import { dataMovies, selectValues } from "../interfaces/Movies";
 import { moviesData } from "../movies.js";
 import Filter from "./movie/Filter";
 import Movie from "./movie/Movie.js";
+import Pagination from "./movie/Pagination";
 
 const ListMovies = () => {
   const [isLoading, setIsLoading] = useState(true);
@@ -50,12 +50,6 @@ const ListMovies = () => {
 
   const endOffset = itemOffset + itemsPerPage;
   const paginateMovies = filteredItems.slice(itemOffset, endOffset);
-  const pageCount = Math.ceil(filteredItems.length / itemsPerPage);
-
-  const handlePageClick = (event: any) => {
-    const newOffset = (event.selected * itemsPerPage) % filteredItems.length;
-    setItemOffset(newOffset);
-  };
 
   if (isLoading) {
     return <p>Loading...</p>;
@@ -83,34 +77,12 @@ const ListMovies = () => {
           />
         ))}
       </div>
-      <ReactPaginate
-        breakLabel="..."
-        nextLabel="next >"
-        onPageChange={handlePageClick}
-        pageRangeDisplayed={5}
-        pageCount={pageCount}
-        previousLabel="< previous"
-        className="flex justify-center gap-2 mt-12"
-        previousLinkClassName="rounded-full bg-gray-100 text-sm px-3 h-8 flex justify-center items-center"
-        nextLinkClassName="rounded-full bg-gray-100 text-sm px-3 h-8 flex justify-center items-center"
-        pageLinkClassName="rounded-full bg-gray-100 flex justify-center items-center px-3 h-8 text-sm"
-        activeLinkClassName="bg-gray-200 font-bold"
+      <Pagination
+        itemsPerPage={itemsPerPage}
+        filteredItemsLength={filteredItems.length}
+        setItemOffset={setItemOffset}
+        setItemsPerPage={setItemsPerPage}
       />
-      <div className="flex justify-center items-center mt-4">
-        <label htmlFor="moviesPerPage">Movies per page:</label>
-
-        <select
-          className="ml-2 bg-gray-100 rounded-full h-8 px-2"
-          name="moviesPerPage"
-          id="moviesPerPage"
-          value={itemsPerPage}
-          onChange={(e) => setItemsPerPage(Number(e.target.value))}
-        >
-          <option value={4}>4</option>
-          <option value={8}>8</option>
-          <option value={12}>12</option>
-        </select>
-      </div>
     </div>
   );
 };
